requirements.txt
streamlit>=1.27.0
chromadb>=0.4.18
openai>=1.3.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
nltk>=3.8.1
scikit-learn>=1.3.0
rank-bm25>=0.2.2
numpy>=1.24.0
pydantic>=2.4.2
python-docx>=1.1.0
playwright>=1.30.0
beautifulsoup4>=4.12.2
python-pptx>=0.6.21
pandas>=1.5.0
openpyxl>=3.0.10